requirements.txt
python-dotenv
supabase
streamlit
pandas
scikit-learn
joblib
psycopg2-binary
xlsxwriter
gTTS
openpyxl
